Function now templated on the Module it's constructed from
In order to have Function respect the const-qualification
of the Module it's built on, Function is now a template class.
Function<Module> provides mutable iterators, while
Function<const Module> provides const iterators.
A Function<Module> can be implicitly converted to a Function<const Module>,
but not vice-versa.

In order to burden the user with this as little as possible,
the build_functions factory is now a function instead of a static method